Main+MainMenuScene: move console starting to `MainMenuScene`
This commit does two things: it adds a button to start the console from
the main menu and moves the logic to start the console that was inside
of `main.cpp` to `MainMenuScene.ccp`.
This means that instead of always starting a console when the
application starts up, it is now up to the user to start it if they need
it.