Meta: switch coding style to the one set by Qt
This change updates the current state of the project to use a unified,
Qt-like coding style. The `.clang-format` file was taken as-is from the
Qt5 repo with the exception of the column limit which was set to 200.
I don't believe this project will need to be coded on small enough
screens that essentially removing the column limit is an issue. Lines
often reach the limit of 100 and have to be wrapped to fit the limit
even when going 5 or 10 characters over the limit. Most screen nowadays
can definitely fit longer lines of text and I believe the limit of 200
is appropriate for a modern project. The limit is still in place to
avoid incredibly long lines. If a line of code has to be over 200
characters, it might be worth breaking it up or most likely completely
rethinking it.

Another notable difference from the Qt code style is the use of prefixes
`m`, `s`, and `g` at the front of private data members, static data
members, and global variables.

Finally, notable liberties taken since there aren't any set rules in the
Qt standard regarding them are:
  - the use of east const,
  - placing slots after signals, and both after other access specifiers.